requirements.txt
vt-py
infi.systray